requirements.txt
streamlit>=1.41.1
google.generativeai
sentence-transformers
# pysqlite3-binary
python-dotenv
Pillow
PyPDF2